Refactor: Convert memo_tool to use register_tools pattern for consistency
